requirements.txt
ipykernel
langchain_core
python-dotenv
streamlit
langchain_community
langserve
langchain
bs4
pypdf
chromadb